+++ internal/admission/validator.go
@@ -297,9 +297,6 @@ func (validator KongHTTPValidator) ValidatePlugin(
 		return false, ErrTextPluginConfigInvalid, err
 	}
 	if k8sPlugin.ConfigFrom != nil {
-		if len(plugin.Config) > 0 {
-			return false, ErrTextPluginUsesBothConfigTypes, nil
-		}
 		config, err := kongstate.SecretToConfiguration(validator.SecretGetter, (*k8sPlugin.ConfigFrom).SecretValue, k8sPlugin.Namespace)
 		if err != nil {
 			return false, ErrTextPluginSecretConfigUnretrievable, err