+++ internal/admission/validator_test.go
@@ -152,27 +152,6 @@ func TestKongHTTPValidator_ValidatePlugin(t *testing.T) {
 			wantErr:     true,
 		},
 		{
-			name:      "plugin has both Config and ConfigFrom",
-			PluginSvc: &fakePluginSvc{},
-			args: args{
-				plugin: kongv1.KongPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names": "whatever"}`),
-					},
-					ConfigFrom: &kongv1.ConfigSource{
-						SecretValue: kongv1.SecretValueFromSource{
-							Key:    "key-auth-config",
-							Secret: "conf-secret",
-						},
-					},
-				},
-			},
-			wantOK:      false,
-			wantMessage: ErrTextPluginUsesBothConfigTypes,
-			wantErr:     false,
-		},
-		{
 			name:      "plugin ConfigFrom references non-existent Secret",
 			PluginSvc: &fakePluginSvc{},
 			args: args{
@@ -282,28 +261,6 @@ func TestKongHTTPValidator_ValidateClusterPlugin(t *testing.T) {
 			wantOK:      false,
 			wantMessage: ErrTextPluginConfigInvalid,
 			wantErr:     true,
-		},
-		{
-			name:      "plugin has both Config and ConfigFrom",
-			PluginSvc: &fakePluginSvc{},
-			args: args{
-				plugin: kongv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names": "whatever"}`),
-					},
-					ConfigFrom: &kongv1.NamespacedConfigSource{
-						SecretValue: kongv1.NamespacedSecretValueFromSource{
-							Key:       "key-auth-config",
-							Secret:    "conf-secret",
-							Namespace: "default",
-						},
-					},
-				},
-			},
-			wantOK:      false,
-			wantMessage: ErrTextPluginUsesBothConfigTypes,
-			wantErr:     false,
 		},
 		{
 			name:      "plugin ConfigFrom references non-existent Secret",